moderator-worker/requirements.txt
redis==5.0.1
requests==2.31.0
opencv-python-headless==4.8.1.78
torch==2.1.0
transformers==4.35.2
Pillow==10.1.0
numpy==1.24.3
boto3==1.34.0

